+++ src/output/saveFileManager/SaveFileManager.cpp
@@ -117,7 +117,7 @@ void SaveFileManager::OnSaveFileCreated()
     ClearEventFlag(static_cast<EventFlags>(FLAG_MILES_MR_WINDYSTONE));
     ClearEventFlag(static_cast<EventFlags>(FLAG_E102_MR_ENTRANCE_WINDY));
     ClearEventFlag(static_cast<EventFlags>(FLAG_E102_MR_WINDYSTONE));
-    
+
     //Ice Cap Stone
     ClearEventFlag(static_cast<EventFlags>(FLAG_SONIC_SS_ICESTONE));
     ClearEventFlag(static_cast<EventFlags>(FLAG_MILES_SS_ICESTONE));
@@ -175,7 +175,7 @@ void SaveFileManager::OnSaveFileCreated()
     ClearEventFlag(static_cast<EventFlags>(FLAG_GET_GOLDEGG));
     ClearEventFlag(static_cast<EventFlags>(FLAG_GET_SILVEREGG));
     ClearEventFlag(static_cast<EventFlags>(FLAG_GET_BLACKEGG));
-    
+
     ClearEventFlag(static_cast<EventFlags>(FLAG_SONIC_ARRIVE_IN_SS));
     ClearEventFlag(static_cast<EventFlags>(FLAG_SONIC_ARRIVE_IN_MR));
     ClearEventFlag(static_cast<EventFlags>(FLAG_SONIC_ARRIVE_IN_EC));
@@ -194,5 +194,27 @@ void SaveFileManager::OnSaveFileCreated()
     ClearEventFlag(static_cast<EventFlags>(FLAG_BIG_ARRIVE_IN_SS));
     ClearEventFlag(static_cast<EventFlags>(FLAG_BIG_ARRIVE_IN_MR));
     ClearEventFlag(static_cast<EventFlags>(FLAG_BIG_ARRIVE_IN_EC));
+    WriteSaveFile();
+}
+
+
+void SaveFileManager::StartAllMissions()
+{
+    for (int i = 0; i < 60; i++)
+    {
+        if (!(MissionFlags[i] & MissionFlags_Complete))
+        {
+            MissionFlags[i] |= MissionFlags_Found;
+            MissionFlags[i] |= MissionFlags_Started;
+        }
+    }
+    WriteSaveFile();
+}
+
+void SaveFileManager::SetMissionCompleted(const int mission)
+{
+    MissionFlags[mission - 1] |= MissionFlags_Found;
+    MissionFlags[mission - 1] &= ~MissionFlags_Started;
+    MissionFlags[mission - 1] |= MissionFlags_Complete;
     WriteSaveFile();
 }