+++ src/output/saveFileManager/SaveFileManager.h
@@ -5,6 +5,9 @@ class SaveFileManager
 public:
     void OnSaveFileLoaded();
     void OnFrame();
+    void StartAllMissions();
+    void SetMissionCompleted(int mission);
+
 private:    
     void OnSaveFileCreated();
 };